Fix signup test
