+++ lib/msal-custom-auth/src/core/network_client/types/ApiErrorResponseTypes.ts
@@ -3,64 +3,42 @@
  * Licensed under the MIT License.
  */
 
-/**
- * Enum for possible error types in Microsoft Entra signup
- */
-export enum SignupErrorType {
-    INVALID_REQUEST = "invalid_request",
-    INVALID_CLIENT = "invalid_client",
-    UNAUTHORIZED_CLIENT = "unauthorized_client",
-    UNSUPPORTED_CHALLENGE_TYPE = "unsupported_challenge_type",
-    USER_ALREADY_EXISTS = "user_already_exists",
-    INVALID_GRANT = "invalid_grant",
-}
+export const CustomAuthApiErrorCode = {
+    CONTINUATION_TOKEN_MISSING: "continuation_token_missing",
+    INVALID_RESPONSE_BODY: "invalid_response_body",
+    EMPTY_RESPONSE: "empty_response",
+    UNSUPPORTED_CHALLENGE_TYPE: "unsupported_challenge_type",
+    ACCESS_TOKEN_MISSING: "access_token_missing",
+    ID_TOKEN_MISSING: "id_token_missing",
+    REFRESH_TOKEN_MISSING: "refresh_token_missing",
+    INVALID_EXPIRES_IN: "invalid_expires_in",
+    INVALID_TOKEN_TYPE: "invalid_token_type",
+    HTTP_REQUEST_FAILED: "http_request_failed",
+    INVALID_REQUEST: "invalid_request",
+    USER_NOT_FOUND: "user_not_found",
+    INVALID_GRANT: "invalid_grant",
+    CREDENTIAL_REQUIRED: "credential_required",
+    ATTRIBUTES_REQUIRED: "attributes_required",
+    USER_ALREADY_EXISTS: "user_already_exists",
+};
 
-export enum SignupErrorChallengeType {
-    INVALID_REQUEST = "invalid_request",
-    EXPIRED_TOKEN = "expired_token",
-    UNSUPPORTED_CHALLENGE_TYPE = "unsupported_challenge_type",
-    INVALID_GRANT = "invalid_grant",
-}
-
-export enum SignupContinueErrorType {
-    INVALID_REQUEST = "invalid_request",
-    EXPIRED_TOKEN = "expired_token",
-    CREDENTIAL_REQUIRED = "credential_required",
-    INVALID_GRANT = "invalid_grant",
-}
+export const CustomAuthApiSuberror = {
+    PASSWORD_TOO_WEAK: "password_too_weak",
+    PASSWORD_TOO_SHORT: "password_too_short",
+    PASSWORD_TOO_LONG: "password_too_long",
+    PASSWORD_RECENTLY_USED: "password_recently_used",
+    PASSWORD_BANNED: "password_banned",
+    PASSWORD_IS_INVALID: "password_is_invalid",
+    INVALID_OOB_VALUE: "invalid_oob_value",
+    ATTRIBUTE_VALIATION_FAILED: "attribute_validation_failed",
+    NATIVEAUTHAPI_DISABLED: "nativeauthapi_disabled",
+};
 
 export enum GenericErrorType {
     UNKNOWN_ERROR = "unknown_error",
     INVALID_JSON_RESPONSE = "invalid_json_response",
 }
 
-export type ApiFieldErrorType = SignupErrorType | SignupErrorChallengeType | SignupContinueErrorType | GenericErrorType;
-/**
- * Enum for invalid client suberrors
- */
-export enum InvalidClientSuberror {
-    NATIVEAUTHAPI_DISABLED = "nativeauthapi_disabled",
-}
-
-export enum InvalidContinueSuberror {
-    INVALID_OOB_VALUE = "invalid_oob_value",
-}
-
-/**
- * Enum for invalid grant suberrors (password-related)
- */
-export enum InvalidGrantSuberror {
-    PASSWORD_TOO_WEAK = "password_too_weak",
-    PASSWORD_TOO_SHORT = "password_too_short",
-    PASSWORD_TOO_LONG = "password_too_long",
-    PASSWORD_RECENTLY_USED = "password_recently_used",
-    PASSWORD_BANNED = "password_banned",
-    PASSWORD_IS_INVALID = "password_is_invalid",
-}
-
-/**
- * Interface for invalid attribute details
- */
 export interface InvalidAttribute {
     name: string;
     reason: string;
@@ -70,12 +48,12 @@ export interface InvalidAttribute {
  * Detailed error interface for Microsoft Entra signup errors
  */
 export interface ApiErrorResponse {
-    error: ApiFieldErrorType;
+    error: string;
     error_description: string;
     error_codes: number[];
     timestamp: string;
     trace_id: string;
     correlation_id: string;
-    suberror?: InvalidClientSuberror | InvalidGrantSuberror | InvalidContinueSuberror;
+    suberror?: string;
     invalid_attributes?: InvalidAttribute[];
 }
